requirements.txt
boto3
psycopg2-binary==2.9.7
